+++ z2/src/bin/z2.rs
@@ -112,8 +112,18 @@ struct ConvertConfigStruct {
     zq2_config_file: String,
     #[arg(value_parser = SecretKey::from_hex)]
     secret_key: SecretKey,
+    #[clap(flatten)]
+    convert_type_group: ConvertTypeGroup,
+}
+#[derive(Args, Debug)]
+#[group(required = true, multiple = false)]
+pub struct ConvertTypeGroup {
     #[clap(long)]
-    skip_accounts: bool,
+    #[arg(default_value_t = false)]
+    convert_accounts: bool,
+    #[clap(long)]
+    #[arg(default_value_t = false)]
+    convert_blocks: bool,
 }
 
 #[derive(Args, Debug)]
@@ -521,7 +531,8 @@ async fn main() -> Result<()> {
                     &arg.zq2_data_dir,
                     &arg.zq2_config_file,
                     arg.secret_key,
-                    arg.skip_accounts,
+                    arg.convert_type_group.convert_accounts,
+                    arg.convert_type_group.convert_blocks,
                 )
                 .await?;
                 Ok(())